requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
transformers
sentence-transformers
torch